feat: change size of `CreditLineBalance` struct fields

+++ src/common/interfaces/ILiquidityPoolAccountable.sol
@@ -14,8 +14,8 @@ interface ILiquidityPoolAccountable is ILiquidityPool {
 
     /// @dev A struct that defines the credit line balance.
     struct CreditLineBalance {
-        uint128 borrowable; // The amount of tokens that can be borrowed.
-        uint128 addons;     // The amount of tokens that are locked as addons.
+        uint64 borrowable; // The amount of tokens that can be borrowed.
+        uint64 addons;     // The amount of tokens that are locked as addons.
     }
 
     // -------------------------------------------- //
